Move the .tsx files in /components/layout and /components/modal folders into /_component folders within /app
