+++ ui/src/app/_components/layout/navbar/MobileNavbar.tsx
