+++ ui/tests/app/_components/layout/footer/Footer.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import Footer from '@/components/layout/footer/Footer';
+import Footer from '@/app/_components/layout/footer/Footer';
 
 describe ('Footer', () => {
