+++ ui/tests/app/_components/layout/navbar/LoginButton.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import Login from '@/components/layout/navbar/LoginButton';
+import Login from '@/app/_components/layout/navbar/LoginButton';
 import { redirect } from 'next/navigation';
 import User, { AnonymousUser } from '@/access/User';
 import Role from '@/access/Role';
